pyproject.toml
[tool.poetry]
name = "stock-exchange-bot"
version = "0.1.0"
description = "Stock exchange bot which performs simple strategy"
authors = ["kndrvt"]

[tool.poetry.dependencies]
python = "3.8.5"
docker = "^4.4.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"